chore: code refactor exlm-1140
