+++ blocks/browse-filters/browse-filter-utils.js
@@ -306,3 +306,12 @@ export function showSearchSuggestionsOnInputClick() {
   }
   toggleSearchSuggestionsVisibility(true);
 }
+
+export const handleCoverSearchSubmit = (targetSearchText) => {
+  const [currentSearchString] = window.location.hash.match(/\bq=([^&#]*)/) || [];
+  if (currentSearchString) {
+    window.location.hash = window.location.hash.replace(currentSearchString, `q=${targetSearchText || ''}`);
+  } else {
+    window.location.hash = `#q=${targetSearchText || ''}&${window.location.hash.slice(1)}`;
+  }
+};
