+++ scripts/coveo-headless/engine.js
@@ -1,10 +1,10 @@
-import { generateCustomContext, generateMlParameters } from '../coveo-utils.js';
+import { generateCustomContext, generateMlParameters } from '../search/search-utils.js';
 
 export default function buildHeadlessSearchEngine(module, coveoToken) {
   return module.buildSearchEngine({
     configuration: {
-      organizationId: 'adobev2prod9e382h1q',
-      organizationEndpoints: module.getOrganizationEndpoints('adobev2prod9e382h1q'),
+      organizationId: 'adobesystemsincorporatednonprod1', // 'adobev2prod9e382h1q',
+      organizationEndpoints: module.getOrganizationEndpoints('adobesystemsincorporatednonprod1'), // ('adobev2prod9e382h1q'),
       accessToken: coveoToken,
       preprocessRequest: (request, clientOrigin, metadata) => {
         const { body } = request;
